# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.DeviceResponse:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      device_token:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      device_type:
        example: desktop
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      last_seen_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      name:
        example: My Desktop
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  handlers.ListDevicesResponse:
    properties:
      devices:
        items:
          $ref: '#/definitions/handlers.DeviceResponse'
        type: array
      total:
        example: 5
        type: integer
    type: object
...
  handlers.RegisterDeviceRequest:
    properties:
      device_type:
        enum:
        - desktop
        - mobile
        example: desktop
        type: string
      name:
        example: My Desktop
        type: string
    required:
    - device_type
    - name
    type: object
  handlers.RegisterDeviceResponse:
    properties:
      device:
        $ref: '#/definitions/handlers.DeviceResponse'
      device_token:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  handlers.UpdateDeviceRequest:
    properties:
      device_type:
        enum:
        - desktop
        - mobile
        example: desktop
        type: string
      name:
        example: My Updated Desktop
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@Backend.app
    name: Backend Support
    url: https://Backend.app/support
  description: Backend - SaaS приложение для файлообмена с акцентом на максимальную скорость
    передачи через Local Network First подход
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://Backend.app/terms
  title: Backend API
  version: "1.0"
paths:
...
  /devices:
    get:
      consumes:
      - application/json
      description: Возвращает список всех устройств пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Список устройств
          schema:
            $ref: '#/definitions/handlers.ListDevicesResponse'
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Список устройств
      tags:
      - devices
    post:
      consumes:
      - application/json
      description: Регистрирует новое устройство для пользователя и возвращает device_token
        для QR-кода
      parameters:
      - description: Данные устройства
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterDeviceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Устройство успешно зарегистрировано
          schema:
            $ref: '#/definitions/handlers.RegisterDeviceResponse'
        "400":
          description: Неверный формат данных
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Регистрация устройства
      tags:
      - devices
  /devices/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет устройство пользователя
      parameters:
      - description: ID устройства
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Устройство удалено
          schema:
            additionalProperties:
              type: boolean
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа к устройству
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Устройство не найдено
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление устройства
      tags:
      - devices
    get:
      consumes:
      - application/json
      description: Возвращает информацию об устройстве по ID
      parameters:
      - description: ID устройства
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация об устройстве
          schema:
            $ref: '#/definitions/handlers.DeviceResponse'
        "400":
          description: Неверный ID устройства
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа к устройству
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Устройство не найдено
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение устройства
      tags:
      - devices
    put:
      consumes:
      - application/json
      description: Обновляет информацию об устройстве
      parameters:
      - description: ID устройства
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Устройство обновлено
          schema:
            $ref: '#/definitions/handlers.DeviceResponse'
        "400":
          description: Неверный формат данных
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа к устройству
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Устройство не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновление устройства
      tags:
      - devices
  /devices/{id}/last-seen:
    post:
      consumes:
      - application/json
      description: Обновляет время последней активности устройства (не требует аутентификации)
      parameters:
      - description: ID устройства
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Время обновлено
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Неверный ID устройства
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Устройство не найдено
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновление времени последней активности
      tags:
      - devices
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"